Implement multiple operation expressions
You can now press an operator after inputting an expression to
immediately solve the expression and use the answer from that
expression as the first operand of the new expression.

The equal button now also checks that you have a full expression and
resets the expression after solving it. This prevents users from
pressing the button multiple times in a row and seeing incorrect
statements in the display.